puntos

+++ Userland/SampleCodeModule/snakeModule.c
@@ -65,7 +65,7 @@ int playSnk(Fruit fruit, Snake snake) {
 	int playing = 1;
 	int exitStatus = 0;
 	while (playing) {
-		printPts(getSecond());
+		printPts((int)getSecond());
 		wait(15);
 		char command = getChar();
 		if (command == '\b') {
